refactor(api): api 폴더 경로 변경
src 폴더 하위에 두고 모아서 관리하기로 결정함
