+++ card-capture/src/api/index.ts
@@ -1,0 +1,5 @@
+import imageApi from './imageApi';
+import loginApi from '@/api/loginApi';
+import promptApi from '@/api/promptApi';
+
+export { imageApi, loginApi, promptApi };
