--- card-capture/src/app/prompt/api/index.ts
@@ -1,3 +1,0 @@
-import promptApi from './promptApi';
-
-export { promptApi };
