--- card-capture/src/components/common/Login/api/index.ts
@@ -1,3 +1,0 @@
-import loginApi from './loginApi';
-
-export { loginApi };
