--- card-capture/src/components/editor/Tab/api/index.ts
@@ -1,3 +1,0 @@
-import imageApi from './imageApi';
-
-export { imageApi };
